Add Tealdeer
I hope I may shamelessly ask to put Tealdeer on the tldr website...

It's quite mature by now, so I released version 1.0.0 yesterday. It
supports caching, conforms to the XDG basedir specification for cached
files, implements both the old and the new tldr-pages format, uses
almost the same command line API as the NodeJS client and it's the
fastest of all clients I tested:
https://github.com/dbrgn/tealdeer/#goals

There is a package for Arch Linux, otherwise users currently need to use
cargo to install the package. But binary releases are planned.